% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_batplot.R
\name{complete_weeks}
\alias{complete_weeks}
\title{Count the weeks and complete them}
\usage{
complete_weeks(count_day_spp_df)
}
\arguments{
\item{count_day_spp_df}{output of count_day_species}
}
\value{
data frame with completed weeks
}
\description{
Count the weeks and complete them
}
