% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_observation_dl_process.R
\name{convert_julian_to_gantt}
\alias{convert_julian_to_gantt}
\title{fill in a gantt chart}
\usage{
convert_julian_to_gantt(.top_10_julian_day = top_10_julian_day)
}
\description{
fill in a gantt chart
}
