% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_batplot.R
\name{count_day_species}
\alias{count_day_species}
\title{count species per day}
\usage{
count_day_species(selected_site_result)
}
\arguments{
\item{selected_site_result}{the output of mapselector::get_subset_site}
}
\value{
data frame with days as data
}
\description{
count species per day
}
