% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_batplot.R
\name{df_to_plot}
\alias{df_to_plot}
\title{take a bat df and make a bat plot}
\usage{
df_to_plot(df)
}
\arguments{
\item{df}{data frame as returned by mapselector::get_subset_site}
}
\value{

}
\description{
take a bat df and make a bat plot
}
