% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_observation_dl_process.R
\name{filter_julday_by_top}
\alias{filter_julday_by_top}
\title{take only the top 10 species in each site}
\usage{
filter_julday_by_top(count_taxa_df = count_taxa, top_spp = top10species)
}
\description{
take only the top 10 species in each site
}
