% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_observation_dl_process.R
\name{filter_plot_gantt}
\alias{filter_plot_gantt}
\title{function that takes a site name and spits out a plot}
\usage{
filter_plot_gantt(site_selected, gantt_df)
}
\description{
function that takes a site name and spits out a plot
}
