% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_observation_dl_process.R
\name{plot_some_bats}
\alias{plot_some_bats}
\title{function to make filled area figure for bats}
\usage{
plot_some_bats(bat_df)
}
\arguments{
\item{bat_df}{data frame for making the area figure}
}
\value{

}
\description{
function to make filled area figure for bats
}
