% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_observation_dl_process.R
\name{process_plot_bats}
\alias{process_plot_bats}
\title{function that takes a data frame and spits out a plot}
\usage{
process_plot_bats(bat_df)
}
\description{
function that takes a data frame and spits out a plot
}
